+++ changelog.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Convex.Utxos.singleton`
   - `Convex.Utxos`: Added redeemer and ex units to `RemoveUtxoEvent`
 * `convex-devnet`: Some haddocks in `Convex.Devnet.NodeQueries`
+* `convex-mockchain`: 
+  - `Convex.MockChain`: Support for profiling plutus scripts. `evaluateTx` returns the script contexts for a transaction. These can be turned into a fully applied script with `fullyAppliedScript`
 
 ## [0.0.1] - 2023-04-26